move All SkinCore Listener code to a separate file
