+++ sdk/react/ooyalaSkinCore.js
@@ -32,46 +32,21 @@ var {
 } = Constants;
 var VideoView = require('./videoView');
 var LanguageSelectionPanel = require('./languageSelectionPanel.js');
+var OoyalaSkinBridgeListener = require('./ooyalaSkinBridgeListener');
 var previousScreenType;
 
 var OoyalaSkinCore = function(ooyalaSkin, eventBridge) {
   this.skin = ooyalaSkin;
   this.bridge = eventBridge;
+  this.ooyalaSkinBridgeListener = new OoyalaSkinBridgeListener(ooyalaSkin, this, eventBridge);
 };
 
 OoyalaSkinCore.prototype.mount = function(eventEmitter) {
-  this.listeners = [];
-  var listenerDefinitions = [
-    [ 'timeChanged',              (event) => this.onTimeChange(event) ],
-    [ 'currentItemChanged',       (event) => this.onCurrentItemChange(event) ],
-    [ 'frameChanged',             (event) => this.onFrameChange(event) ],
-    [ 'volumeChanged',            (event) => this.onVolumeChanged(event) ],
-    [ 'playCompleted',            (event) => this.onPlayComplete(event) ],
-    [ 'stateChanged',             (event) => this.onStateChange(event) ],
-    [ 'discoveryResultsReceived', (event) => this.onDiscoveryResult(event) ],
-    [ 'onClosedCaptionUpdate',    (event) => this.onClosedCaptionUpdate(event) ],
-    [ 'adStarted',                (event) => this.onAdStarted(event) ],
-    [ 'adSwitched',               (event) => this.onAdSwitched(event) ],
-    [ 'adPodCompleted',           (event) => this.onAdPodCompleted(event) ],
-    [ 'setNextVideo',             (event) => this.onSetNextVideo(event) ],
-    [ 'upNextDismissed',          (event) => this.onUpNextDismissed(event) ],
-    [ 'playStarted',              (event) => this.onPlayStarted(event) ],
-    [ 'postShareAlert',           (event) => this.onPostShareAlert(event) ],
-    [ 'error',                    (event) => this.onError(event) ],
-    [ 'embedCodeSet',             (event) => this.onEmbedCodeSet(event) ]
-  ];
-
-  for (var i = 0; i < listenerDefinitions.length; i++) {
-    var d = listenerDefinitions[i];
-    this.listeners.push(eventEmitter.addListener( d[0], d[1] ) );
-  }
+  this.ooyalaSkinBridgeListener.mount(eventEmitter);
 };
 
 OoyalaSkinCore.prototype.unmount = function() {
-  for (var i = 0; i < this.listeners.length; i++) {
-    this.listeners[i].remove;
-  }
-  this.listeners = [];
+  this.ooyalaSkinBridgeListener.unmount();
 };
 
 // event handlers.
@@ -106,6 +81,19 @@ OoyalaSkinCore.prototype.onOptionDismissed = function() {
   this.popFromOverlayStackAndMaybeResume();
 };
 
+OoyalaSkinCore.prototype.onDiscoveryRow = function(info) {
+  this.bridge.onDiscoveryRow(info);
+};
+
+OoyalaSkinCore.prototype.onOverlayDismissed = function() {
+  Log.log("On Overlay Dismissed");
+  this.popFromOverlayStackAndMaybeResume();
+}
+
+OoyalaSkinCore.prototype.onLanguageSelected = function(e) {
+  Log.log('onLanguageSelected:'+e);
+  this.skin.setState({selectedLanguage:e});
+};
 /**
  *  When a button is pressed on the control bar
  *  If it's a "fast-access" options button, open options menu and perform the options action
@@ -151,163 +139,8 @@ OoyalaSkinCore.prototype.updateClosedCaptions = function() {
   }
 };
 
-OoyalaSkinCore.prototype.onClosedCaptionUpdate = function(e) {
-  this.skin.setState( {captionJSON: e} );
-};
-
-OoyalaSkinCore.prototype.onDiscoveryRow = function(info) {
-  this.bridge.onDiscoveryRow(info);
-};
-
-OoyalaSkinCore.prototype.onTimeChange = function(e) { // todo: naming consistency? playheadUpdate vs. onTimeChange vs. ...
-  this.skin.setState({
-    playhead: e.playhead,
-    duration: e.duration,
-    initialPlay: false,
-    availableClosedCaptionsLanguages: e.availableClosedCaptionsLanguages,
-    cuePoints: e.cuePoints,
-  });
-
-  if(this.skin.state.screenType == SCREEN_TYPES.VIDEO_SCREEN ||
-     this.skin.state.screenType == SCREEN_TYPES.END_SCREEN) {
-    this.previousScreenType = this.skin.state.screenType;
-  }
-  this.updateClosedCaptions();
-};
-
-OoyalaSkinCore.prototype.onAdStarted = function(e) {
-  Log.log( "onAdStarted");
-  Log.log(e);
-  this.skin.setState({ad:e, screenType:SCREEN_TYPES.VIDEO_SCREEN});
-};
-
-OoyalaSkinCore.prototype.onAdSwitched = function(e) {
-  Log.log( "onAdSwitched");
-  this.skin.setState({ad:e});
-};
-
-OoyalaSkinCore.prototype.onAdPodCompleted = function(e) {
-  Log.log( "onAdPodCompleted ");
-  this.skin.setState({ad: null});
-};
-
-OoyalaSkinCore.prototype.onCurrentItemChange = function(e) {
-  Log.log("currentItemChangeReceived, promoUrl is " + e.promoUrl);
-  this.skin.setState({
-    title:e.title,
-    description:e.description,
-    duration:e.duration,
-    live:e.live,
-    promoUrl:e.promoUrl,
-    hostedAtUrl: e.hostedAtUrl,
-    playhead:e.playhead,
-    width:e.width,
-    height:e.height,
-    volume:e.volume,
-    captionJSON:null});
-  if (!this.skin.state.autoPlay) {
-    this.skin.setState({screenType: SCREEN_TYPES.START_SCREEN});
-  };
-  this.clearOverlayStack();
-};
-
-OoyalaSkinCore.prototype.onFrameChange = function(e) {
-  Log.log("receive frameChange, frame width is" + e.width + " height is" + e.height);
-  this.skin.setState({width:e.width, height:e.height, fullscreen:e.fullscreen});
-};
-
-OoyalaSkinCore.prototype.onPlayStarted = function(e) {
-  Log.log("Play Started received");
-  this.skin.setState({screenType: SCREEN_TYPES.VIDEO_SCREEN, autoPlay: false});
-};
-
-OoyalaSkinCore.prototype.onPlayComplete = function(e) {
-  Log.log("Play Complete received: upNext dismissed: "  + this.skin.state.upNextDismissed);
-  if (this.shouldShowDiscoveryEndscreen()) {
-      this.pushToOverlayStackAndMaybePause(OVERLAY_TYPES.DISCOVERY_SCREEN);
-  }
-  this.skin.setState({screenType: SCREEN_TYPES.END_SCREEN});
-};
-
-OoyalaSkinCore.prototype.onOverlayDismissed = function() {
-  Log.log("On Overlay Dismissed");
-  this.popFromOverlayStackAndMaybeResume();
-}
-
-OoyalaSkinCore.prototype.onDiscoveryResult = function(e) {
-  Log.log("onDiscoveryResult results are:", e.results);
-  this.skin.setState({discoveryResults:e.results});
-  if(e.results) {
-    this.onSetNextVideo({nextVideo:e.results[0]})
-  }
-};
-
-OoyalaSkinCore.prototype.onStateChange = function(e) {
-  Log.log("state changed to:" + e.state)
-  switch (e.state) {
-    case "completed":
-    case "error":
-    case "init":
-    case "paused":
-    case "ready":
-      this.skin.setState({
-        playing: false,
-        loading: false
-      });
-      break;
-    case "playing":
-      this.skin.setState({
-        playing: true,
-        loading: false,
-        initialPlay: (this.skin.state.screenType == SCREEN_TYPES.START_SCREEN) ? true : false,
-        screenType: SCREEN_TYPES.VIDEO_SCREEN});
-      break;
-    case "loading":
-      this.skin.setState({
-        loading: true
-      })
-      break;
-    default:
-      break;
-  }
-};
-
-OoyalaSkinCore.prototype.onError = function(e) {
-  Log.log("Error received");
-  this.skin.setState({screenType:SCREEN_TYPES.ERROR_SCREEN, error:e});
-};
-
-OoyalaSkinCore.prototype.onEmbedCodeSet = function(e) {
-  Log.log("EmbedCodeSet received");
-  this.skin.setState({screenType:SCREEN_TYPES.LOADING_SCREEN});
-};
-
-OoyalaSkinCore.prototype.onUpNextDismissed = function(e) {
-  Log.log("SetNextVideo received");
-  this.skin.setState({upNextDismissed:e.upNextDismissed});
-};
-
-OoyalaSkinCore.prototype.onSetNextVideo = function(e) {
-  Log.log("SetNextVideo received");
-  this.skin.setState({nextVideo:e.nextVideo});
-};
-
-OoyalaSkinCore.prototype.onLanguageSelected = function(e) {
-  Log.log('onLanguageSelected:'+e);
-  this.skin.setState({selectedLanguage:e});
-};
-
 OoyalaSkinCore.prototype.shouldShowLandscape = function() {
   return this.skin.state.width > this.skin.state.height;
-};
-
-OoyalaSkinCore.prototype.onPostShareAlert = function(e) {
-  this.skin.setState({alertTitle: e.title});
-  this.skin.setState({alertMessage: e.message});
-};
-
-OoyalaSkinCore.prototype.onVolumeChanged = function(e) {
-  this.skin.setState({volume: e.volume});
 };
 
 OoyalaSkinCore.prototype.renderStartScreen = function() {
